Restructure code: move src/* to root, separate DB config.
