+++ graphql/mutations/session.js
@@ -3,7 +3,7 @@ import {
 } from 'graphql'
 
 import sessionType from '../types/SessionType'
-import Session from '../../model/sessions'
+import Session from '../../model/session/session'
 
 var SessionModel = new Session()
