+++ graphql/types/UserType.js
@@ -5,7 +5,7 @@ import {
 } from 'graphql'
 
 import sessionType from '../types/SessionType'
-import Session from '../../model/sessions'
+import Session from '../../model/session/session'
 
 var SessionModel = new Session()
