+++ gulpfile.babel.js
@@ -6,7 +6,7 @@ import gutil from 'gulp-util'
 gulp.task('build', function (done) {
   process.env.NODE_ENV = yargs.argv.env || 'development'
 
-  var config = require(`${__dirname}/src/config/${process.env.NODE_ENV}/webpack.js`)
+  var config = require(`${__dirname}/config/${process.env.NODE_ENV}/webpack.js`)
   webpack(config, function (err, stats) {
     if (err) {
       throw new gutil.PluginError('webpack', err)
