+++ lib/auth.js
@@ -1,7 +1,7 @@
 import passport from 'passport'
 import { Strategy as LocalStrategy } from 'passport-local'
 
-import User from '../model/users'
+import User from '../model/user/user'
 let userModel = new User()
 
 passport.use(new LocalStrategy({usernameField: 'email'}, (username, password, done) => {
