+++ model/session/session.js
@@ -1,4 +1,4 @@
-import db from './db'
+import db from '../../db/config'
 
 class Session {
