+++ server.js
@@ -1,10 +1,10 @@
 import express from 'express'
 import expressGraphQL from 'express-graphql'
-import schema from './data/schema'
 import jwt from 'jsonwebtoken'
 import ExpressJwt from 'express-jwt'
 import passport from 'passport'
-import authConfig from './config/auth'
+import schema from './graphql/schema'
+import authConfig from './lib/auth'
 
 const server = global.server = express()
